Extracted some interfaces and utils.
